Move record containers to their own directory.
