+++ scripts/containers/record/AddFormPage.js
@@ -2,15 +2,16 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { push as updatePath } from "react-router-redux";
 
-import BulkForm from "../components/record/BulkForm";
-import * as CollectionActions from "../actions/collection";
-import * as NotificationsActions from "../actions/notifications";
+import AddForm from "../../components/record/AddForm";
+import * as CollectionActions from "../../actions/collection";
+import * as NotificationsActions from "../../actions/notifications";
 
 
 function mapStateToProps(state) {
   return {
-    collection: state.collection,
     session: state.session,
+    collection: state.collection,
+    record: state.record,
   };
 }
 
@@ -25,4 +26,4 @@ function mapDispatchToProps(dispatch) {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(BulkForm);
+)(AddForm);
