+++ scripts/containers/record/EditFormPage.js
@@ -2,9 +2,9 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { push as updatePath } from "react-router-redux";
 
-import EditForm from "../components/record/EditForm";
-import * as CollectionActions from "../actions/collection";
-import * as NotificationsActions from "../actions/notifications";
+import EditForm from "../../components/record/EditForm";
+import * as CollectionActions from "../../actions/collection";
+import * as NotificationsActions from "../../actions/notifications";
 
 
 function mapStateToProps(state) {
