+++ scripts/routes.js
@@ -22,9 +22,9 @@ import CollectionListPage from "./containers/collection/CollectionListPage";
 import CollectionHistoryPage from "./containers/collection/CollectionHistoryPage";
 import CollectionCreatePage from "./containers/collection/CollectionCreatePage";
 import CollectionEditPage from "./containers/collection/CollectionEditPage";
-import AddFormPage from "./containers/AddFormPage";
-import BulkFormPage from "./containers/BulkFormPage";
-import EditFormPage from "./containers/EditFormPage";
+import AddFormPage from "./containers/record/AddFormPage";
+import BulkFormPage from "./containers/record/BulkFormPage";
+import EditFormPage from "./containers/record/EditFormPage";
 import * as sessionActions from "./actions/session";
 import * as bucketActions from "./actions/bucket";
 import * as collectionActions from "./actions/collection";
